Plugin uninstall issue fix.
git-svn-id: https://plugins.svn.wordpress.org/miniorange-login-with-eve-online-google-facebook/trunk@1995741 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,17 +1,7 @@
 <?php
-require('manage-avatar.php');
 
-$users = get_users( array() );
-foreach ( $users as $user ) {
-	$attachment_id = get_user_meta( $user->ID, 'mo_oauth_avatar_manager_custom_avatar', true );
-	if ( ! empty( $attachment_id ) ) {
-		mo_oauth_avatar_manager_delete_avatar($attachment_id);
-	}
-	delete_user_meta($user->ID, 'user_eveonline_character_name');
-	delete_user_meta($user->ID, 'user_eveonline_corporation_name');
-	delete_user_meta($user->ID, 'user_eveonline_alliance_name');
-	delete_user_meta($user->ID, 'dismissed_wp_pointers');
-}
+if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) 
+    exit();
 
 delete_option('host_name');
 delete_option('mo_oauth_admin_email');
@@ -46,4 +36,5 @@ delete_option('new_registration');
 delete_option('mo_oauth_registration_status');
 delete_option( 'mo_oauth_client_disable_authorization_header' );
 delete_option('mo_oauth_client_show_mo_server_message');
+
 ?>
